Move iterators into iterators package
